COM-3408 : create Attendance composable
